requirements.txt
redis
psycopg2-binary
anthropic
ragas
datasets
requests
openai 
matplotlib 
numpy
streamlit
exa_py